+++ app/models/email_batch.rb
@@ -6,7 +6,7 @@ class EmailBatch < ApplicationRecord
 
   scope(:in_past_week, -> { where("created_at > ?", 7.days.ago) })
 
-  sig { returns(Integer) }
+  sig { returns(Numeric) }
   def self.total_sent_in_past_week
     in_past_week.sum(:no_emails)
   end
